Move EditorSwitcherManager back to editor utils
