+++ spyder/plugins/editor/utils/switcher_manager.py
